Run A Lighter Version Of Smoketests In Production
The automated SMS journey routinely fails, due to issues with notify, this is
not a reflection of system failing but an issue with the API call. This is
giving false warnings about our production systems so this part of the test has
been disabled in the production environment, but will still be used in staging
and when we deploy new code.