Remove automated section of create new env instructions
Deleting the automated instructions as we could not get the script to run all
the way through when tested. Automating this process is a good idea, but we
need to ensure the script will always run without error. At present it
hangs/fails before completing the setup on "New key pair has been generated and
saved in ../govwifi-build". Removing reference to the bootstrap.sh script from
the instructions to avoid confusion. But will leave the bootstrap.sh script
present as a foundation to iterate on.

Also:
Increase number of bytes for ssh key
Add instructions to merge changes on govwifi-build to master before proceeding