Update AWS service discovery resource
This was causing an error when rebuilding the environment from
scratch. The health_check_custom_config is now deprecated.

This will now build new enviroments without the service discovery
health check (we are using the loadbalancer health check and
the discovery service does not allow us to query port 3000 so
it will now fail). Terraform will ignore the configuration in our
old environments.

More information can be found in this Jira ticket:
https://technologyprogramme.atlassian.net/browse/GW-2513